+++ src/Pages/Login.jsx
@@ -1,22 +1,22 @@
 import { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useRecoilValue } from 'recoil';
-import useLogin from '../Hooks/Sign/useLogin';
 import isDesktop from '../Recoil/isDesktop/isDesktop';
-import PCLogin from '../Components/Sign/PCLogin';
+import isLogin from '../Recoil/isLogin/isLogin';
+import PCSign from '../Components/Sign/PCSign.jsx';
 import LoginForm from '../Components/Sign/LoginForm';
 
 const Login = () => {
   const navigate = useNavigate();
   const isPCScreen = useRecoilValue(isDesktop);
-  const { isLoginState } = useLogin();
+  const isLoginState = useRecoilValue(isLogin);
 
   useEffect(() => {
     if (isLoginState) navigate('/home');
     //eslint-disable-next-line
   }, []);
 
-  return isPCScreen ? <PCLogin /> : <LoginForm />;
+  return isPCScreen ? <PCSign /> : <LoginForm />;
 };
 
 export default Login;
